wallet/pyproject.toml
[project]
name = "wallet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "circle-developer-controlled-wallets>=9.2.0",
]
